+++ test/src/command_filtered_test.dart
@@ -1,0 +1,32 @@
+import 'package:flutter_test/flutter_test.dart';
+import 'package:result_command/src/command.dart';
+import 'package:result_dart/result_dart.dart';
+
+void main() {
+  test(
+    'Filter Ok',
+    () async {
+      final command = Command1<int, int>((int value) async {
+        await Future.delayed(const Duration(milliseconds: 100));
+
+        if (value < 0) {
+          return Failure(Exception('Negative value'));
+        }
+
+        return Success(value);
+      });
+
+      final fiteredSuccess = command.filter(
+        0,
+        (value) => value is SuccessCommand<int> ? value.value : null,
+      );
+
+      fiteredSuccess.addListener(expectAsync0(() {
+        expect(fiteredSuccess.value, 2);
+      }));
+      await command.execute(-1);
+      await command.execute(-2);
+      await command.execute(2);
+    },
+  );
+}